refactor components in new folders
