+++ apps/web/src/app/_components/reviews/existing-company-content.tsx
@@ -12,14 +12,14 @@ import { useCustomToast } from "@cooper/ui/hooks/use-custom-toast";
 import { Input } from "@cooper/ui/input";
 import { Label } from "@cooper/ui/label";
 
-import type { RoleRequestType } from "../new-role-dialogue";
+import type { RoleRequestType } from "../roles/new-role-dialogue";
 import { api } from "~/trpc/react";
-import { CompanyCardPreview } from "../../companies/company-card-preview";
-import { FormSection } from "../../form/form-section";
-import LocationBox from "../../location";
-import { industryOptions } from "../../onboarding/constants";
-import { FormLabel } from "../../themed/onboarding/form";
-import FilterBody from "../../filters/filter-body";
+import { CompanyCardPreview } from "../companies/company-card-preview";
+import { FormSection } from "../form/form-section";
+import LocationBox from "../location";
+import { industryOptions } from "../onboarding/constants";
+import { FormLabel } from "../themed/onboarding/form";
+import FilterBody from "../filters/filter-body";
 
 const filter = new Filter();
 const roleSchema = z.object({
