+++ apps/web/src/app/_components/reviews/unused/review-search-bar.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Input } from "../themed/onboarding/input";
+import { Input } from "../../themed/onboarding/input";
 
 interface ReviewSearchBarProps {
   searchTerm: string;
